rename files and move stuff
