+++ scb_move_straight.go
